% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_smallworld.R
\docType{data}
\name{basic_smallworld}
\alias{basic_smallworld}
\title{Toy Small-world Network Data Example}
\format{
A 500 \eqn{\times} 20 continuous response matrix
}
\usage{
data(basic_smallworld)
}
\description{
A toy small-world network example (\emph{n} = 500)
}
\examples{
# Generated using:
# basic_smallworld <- simulate_smallworld(
#  nodes = 20, # 20 nodes in the network
#  density = 0.30, # moderate initial lattice connectivity
#  rewire = 0.20, # 20\% rewiring probability
#  sample_size = 500 # number of cases = 500
# )$data
data("basic_smallworld")

}
\keyword{datasets}
